Make the various waiting mechanisms more robust by bounding the number of concurrent waits.
This is mainly to cope with web interface people who refresh a few times, leading to outstanding waits. Ideally this would be dealt with through a callback from the HTTP server to alert us that the client has given up and gone home; unfortunately WGSI sucks sufficiently hard that there is no such callback.