Start of rewriting pyCURL code to use a fetcher thread.
Since I have more time to do this now, also taking this opportunity to refactor the code to be less confusing.

Up to here I've written the main, very-general multi-fetcher thread and ported the simpler fetchers to use it.

To come: porting the stream-fetcher. As such this is broken right now.