More work on factoring the two interpreters.
Also converted Skywriting away from local references, which makes this easier.

So far they mostly share the exec/spawn-exec code, i.e. that which concerns interacting with the outside. I'd like to make them share the code to do with wrapping up at the end of an interpretation session too -- that which either returns or creates a task descriptor for next time.