Add a job-starting script that can invoke theoretically any executor/handler.
The way this is achieved is a little bit of a mess right now, due to the conflict between the master's desire to assign a name to the new job and the executors' build-task-descriptor routines' desire to know their task ID so they can name their references. Solution at the moment: the master does the last bit of task descriptor building in response to a /job/newsubmit POST. Better solution: either rewrite the executors to deal with the situation that they can't always name their refs or else permit the submitter to assign a name as a seperate RPC to job creation.