Undo the nasty hack to the master
For now I take the approach that all useful external calls don't use the block store, publish references or care about task_id, so I just define those to be bad ideas.

A better solution would be to either:

(1) split the naming process so that a job submitter can seperately request a job ID and submit a job,
(2) implement some appropriate abstraction such that the executors can say "name and publish this reference" and the right thing will happen depending on whether they're running on the submitter or a worker, or
(3) submit an initial "job" to a worker which creates the initial task descriptor; a sort of preheader.

For now these all seem somewhat unpleasant so I take the lazy option