Rewrite SkyPy smart producer to avoid threads.
This is because whilst Stackless Python can deal with threads and coroutines in the same process, SkyPy can't. This leaves us with no reasonable way to deal with non-cooperative code which we'd rather not allow to run forever, like user SkyPy programs.

For now I use the rubbish solution of draining the output buffer whenever we enter write(), effectively cooperative-ising the code without bothering to use coroutines.

This should work alright when remote users subscribe as if the user isn't calling write() there's nothing to see: the main problem will arise if a lot of remote workers try to subscribe in quick succession to the point that the pipe write blocks. In this case their subscribe() call will block indefinitely.

Possible solution: non-blocking write those sorts of RPCs, and rebound a subscribe that would overfill the pipe complaining of oversubscription?

Or, ask SkyPy users to yield from time to time? Both unpleasant.