SkyPy: save coroutines for newly spawned task from the main coro rather than the user coro
It turns out from investigating PyPy issue 689 that coroutines have a
*statically determined* parent to which they will return when they
finish (drop off the bottom of their stack); this is always their
creator. As the user coro was creating the spawnee's coro, it was the
static parent, and theoretically if the spawnee returned from main, it
would return to its creator (actually this never happens, as my
"start_script" thunk sees to it that it actually exits by switching to
the real 'main' coroutine of the current process).

This caused issue 689 to arise, because the creator was in the static
coro chain when the new one was pickled, but also meant that even with
a workaround, we were pointless pickling a task's creator along with
that task. As this forms a chain, a spawn-chain would be recorded in
its entirity. This way the "parent" is "main", which PyPy records as
a dummy value indicating "whatever is main in this process," rather
than the actual stack of coro-main at the time of writing.

In short: this saves a bunch of pointless work at pickle/unpickle
time, and makes simple tasks' coros 3-4x smaller.